+++ lib/src/serializable_cookie.dart
@@ -8,33 +8,29 @@ class SerializableCookie {
     createTimeStamp = (DateTime.now().millisecondsSinceEpoch ~/ 1000).toInt();
   }
 
-  /// Create a instance form Json string.
-  SerializableCookie.fromJson(String value) {
-    final t = value.split(';_crt=');
-    cookie = Cookie.fromSetCookieValue(t[0]);
-    createTimeStamp = int.parse(t[1]);
-  }
-
-  /// Test the  whether this cookie is expired.
-
-  bool isExpired() {
-    final t = DateTime.now();
-    return (cookie.maxAge != null && cookie.maxAge! < 1) ||
-        (cookie.maxAge != null &&
-            (t.millisecondsSinceEpoch ~/ 1000).toInt() - createTimeStamp >=
-                cookie.maxAge!) ||
-        (cookie.expires != null && !cookie.expires!.isAfter(t));
-  }
-
-  /// Serialize the Json string.
-  String toJson() => toString();
-
-  late Cookie cookie;
-
+  final Cookie cookie;
   int createTimeStamp = 0;
 
-  @override
-  String toString() {
-    return cookie.toString() + ';_crt=$createTimeStamp';
+  /// Create an instance from JSON string.
+  factory SerializableCookie.fromJson(String value) {
+    final t = value.split(';_crt=');
+    return SerializableCookie(Cookie.fromSetCookieValue(t[0]))
+      ..createTimeStamp = int.parse(t[1]);
   }
+
+  /// Tells whether this cookie is expired.
+  bool isExpired() {
+    final now = DateTime.now();
+    return (cookie.maxAge != null && cookie.maxAge! < 1) ||
+        (cookie.maxAge != null &&
+            (now.millisecondsSinceEpoch ~/ 1000).toInt() - createTimeStamp >=
+                cookie.maxAge!) ||
+        (cookie.expires != null && !cookie.expires!.isAfter(now));
+  }
+
+  /// Serialize the JSON string.
+  String toJson() => toString();
+
+  @override
+  String toString() => '$cookie;_crt=$createTimeStamp';
 }